backend/src/main/java/com/interview/resource/GlobalExceptionHandler.java
package com.interview.resource;

import com.interview.exception.ValidationException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

import java.util.HashMap;
import java.util.Map;

@ControllerAdvice
public class GlobalExceptionHandler {

    @ExceptionHandler(ValidationException.class)
    public ResponseEntity<Map<String, Object>> handleValidationException(ValidationException e) {
        Map<String, Object> body = new HashMap<>();
        body.put("message", e.getErrors());
        return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
    }
}
